### 2025-02-18 #7788 @editorialbot
[PRE REVIEW]: visual-gestures.js- Lets you control the cursor with hand or finger movements
**Submitting author:** @BodduSriPavan111 (<a href="http://orcid.org/0009-0003-3233-9447">Boddu Sri Pavan</a>)
**Repository:** https://github.com/learn-hunger/visual-gestures
**Branch with paper.md** (empty if default branch):
**Version:** v0.0.5
