### 2025-02-18 #7788 @crvernon
:wave: @BodduSriPavan111 - I believe your paper may not be compiling due to missing three dashes at the end of the header of your paper.  See:

```
bibliography: paper.bib
---
```

You are missing the `---`
