+++ paper.md
@@ -31,6 +31,7 @@ affiliations:
  
 date: 15 February 2025
 bibliography: paper.bib
+---
 
 # Summary
 
