### 2025-02-18 #7788 @BodduSriPavan-111
> 👋 [@BodduSriPavan111](https://github.com/BodduSriPavan111) - I believe your paper may not be compiling due to missing three dashes at the end of the header of your paper. See:
>
> ```
> bibliography: paper.bib
> ---
> ```
>
> You are missing the `---`

Hi @crvernon ,

Thanks for the quick suggestion. I've made the suggested changes and committed them to the main branch. Please let me know if any further updates are required from our side.
