### 2025-07-28 learn-hunger/visual-gestures#4 @boisgera
MediaPipe Hand Landmarker Reference
JOSS Context: https://github.com/openjournals/joss-reviews/issues/8394

Please add a bibliography entry (or at least a hyperlink in the text) for Google's MediaPipe Hand Landmarker from <https://ai.google.dev/edge/mediapipe/solutions/vision/hand_landmarker> in the JOSS paper.

It would also probably be useful to add a bibliography entry for the GHUM paper which is cited in <https://storage.googleapis.com/mediapipe-assets/Model%20Card%20Hand%20Tracking%20(Lite_Full)%20with%20Fairness%20Oct%202021.pdf>?

GHUM & GHUML: Generative 3D Human Shape and
Aiculated Pose Models
Proceedings of the IEEE/CVF Conference on
Computer Vision and Paern Recognition, pages
6184-6193, 2020
