### 2025-07-28 learn-hunger/visual-gestures#5 @boisgera
Running the example
JOSS Context: https://github.com/openjournals/joss-reviews/issues/8394

I was trying to run locally the example found in `example2`, assuming that in this repository, a `npm install` followed by a `npm run dev` would launch it.

Instead I get:

```
💻  npm run dev

> example2@0.0.0 dev
> vite

  VITE v5.3.5  ready in 124 ms

  ➜  Local:   http://localhost:5173/
  ➜  Network: use --host to expose
  ➜  press h + enter to show help
Error: The following dependencies are imported but could not be resolved:

  @learn-hunger/visual-gestures/src/ (imported by /home/boisgera/tmp/visual-gestures/example2/src/index.ts)
  @learn-hunger/visual-gestures/src/app/utilities/vg-constants (imported by /home/boisgera/tmp/visual-gestures/example2/src/events.ts)

Are they installed?
    at file:///home/boisgera/tmp/visual-gestures/example2/node_modules/vite/dist/node/chunks/dep-mCdpKltl.js:50346:15
    at process.processTicksAndRejections (node:internal/process/task_queues:105:5)
    at async file:///home/boisgera/tmp/visual-gestures/example2/node_modules/vite/dist/node/chunks/dep-mCdpKltl.js:49851:26
```

Is there a problem or is there an issue with my process? If it's the latter, please add a `README.md` file in the `example2` directory (by the way, you could probably rename it `example`, or create and `examples` folder and createa  subfolder with a more descriptive name for you example project there).
