### 2025-07-28 learn-hunger/visual-gestures#6 @boisgera
Getting Started
JOSS Context: https://github.com/openjournals/joss-reviews/issues/8394

I have a hard time connecting the second step of the [Getting Started](JOSS Context: https://github.com/openjournals/joss-reviews/issues/8394) section and an existing web site:

  - You mention that `video` should correspond to 'HTMLVideoElement' with live webcam stream, but

  - `handDetector` is not defined either; you mention some mediapipe handLandmarker docs, but it's not immediately obvious that `handDetector` is what is called `handLandmarker` there (I had to read some of your example2 code to understand that).

  - more generally, how do I integrate this js script "into my website" (you use vite into your example, can you maybe detail that a bit more in this case?)

Alternatively, if this is too complex to fit into the README, provide a dead simple but working example in the github repository and link to it in the README (?). See also https://github.com/learn-hunger/visual-gestures/issues/5
