### 2025-07-28 2cf3e87
docs: renamed example2 to example
