+++ .github/workflows/website.yml
@@ -30,6 +30,6 @@ jobs:
           vercel-org-id: ${{ secrets.VERCEL_ORG_ID}} #Required
           vercel-project-id: ${{ secrets.VERCEL_PROJECT_ID}} #Required
           vercel-args: "--prod --confirm"
-          working-directory: ./example2/dist
+          working-directory: ./example/dist
           alias-domains: | #Optional
             visual-gestures-source.vercel.app
