### 2025-07-28 51f487b
Hotfix/nagendra/issue 5/change folder name (#7)
* docs: renamed example2 to example

* docs: renamed folder

---------

Co-authored-by: nagendra-dharmireddi <nagendra.dharmireddi@cognitivebotics.com>
