+++ README.md
@@ -36,51 +36,81 @@ Tailored for different industries such as controlling machinery in factories, na
 
 **1. Install our npm package**
 
->  ```javascript
->  npm install @learn-hunger/visual-gestures
->  ```
+> ```javascript
+> npm install @learn-hunger/visual-gestures
+> ```
 
 **2. Integrate into your existing website**
 
->  ```javascript
->  import { VisualGestures } from "@learn-hunger/visual-gestures/dist/";
+> ```javascript
+> import { VisualGestures } from "@learn-hunger/visual-gestures/dist/";
+> import {
+>   FilesetResolver,
+>   HandLandmarker,
+>   HandLandmarkerResult,
+> } from "@mediapipe/tasks-vision";
+> // create instance of visual-gestures
+> // which accepts optional parameters of container and the landmark to be used as pointer
+> // [Default body and landmark 8 is used respectively]
+> const vg = new VisualGestures();
 >
->  /**
->   *create instance of visual-gestures
->   *which accepts optional parameters of container and the landmark to be used as pointer
->   *[Default body and landmark 8 is used respectively]
->   */
->  const vg = new VisualGestures();
+> // load weights using task vision handlandmarker
+> // one can take refrence from [https://github.com/learn-hunger/visual-gestures/blob/main/example/src/services/handLandmarks.ts]
+> let handDetector: HandLandmarker;
+> export async function loadWeights() {
+>   if (handDetector) {
+>     return handDetector;
+>   }
+>   const vision = await FilesetResolver.forVisionTasks(
+>     "https://cdn.jsdelivr.net/npm/@mediapipe/tasks-vision@latest/wasm",
+>   );
+>   handDetector = await HandLandmarker.createFromOptions(vision, {
+>     baseOptions: {
+>       modelAssetPath: `https://storage.googleapis.com/mediapipe-models/hand_landmarker/hand_landmarker/float16/1/hand_landmarker.task`,
+>       delegate: "GPU",
+>     },
+>     runningMode: "VIDEO",
+>     numHands: 2,
+>   });
 >
->  // get hand landmarks from mediapipe's taskvision
->  // here video corresponds to 'HTMLVideoElement' with live webcam stream
->  const landmarks = handDetector.detectForVideo(video, performance.now());
->  vg.detect(landmarks.landmarks[0], performance.now());
->  // Virtual cursor can be seen once model loading and detection started successfully
->  ```
+>   return handDetector;
+> }
+> // ask for camera permission [https://github.com/learn-hunger/visual-gestures/blob/main/example/src/utils/camera.ts]
+> // then detect hand landmarks [https://github.com/learn-hunger/visual-gestures/blob/main/example/src/services/handLandmarks.ts]
+> // get hand landmarks from mediapipe's taskvision
+> // here video corresponds to 'HTMLVideoElement' with live webcam stream [https://github.com/learn-hunger/visual-gestures/blob/main/example/src/index.ts]
+> const landmarks = handDetector.detectForVideo(video, performance.now());
+> vg.detect(landmarks.landmarks[0], performance.now());
+> // Virtual cursor can be seen once model loading and detection started successfully
+> ```
+>
+> ```
+>
+> ```
 
[…]
