### 2025-07-28 learn-hunger/visual-gestures#6 @pvscreations
@boisgera
Thank you for the feedback! As suggested, I’ve updated the [README.md](https://github.com/learn-hunger/visual-gestures/blob/main/README.md).
To keep things concise given the scope of the project, the README provides a minimal overview, with direct links to relevant code sections and documentation.

We’ve also included a working example within the repository, and the application has been deployed for easier exploration:
 [Live Demo](https://visual-gestures.vercel.app/)
