### 2025-07-29 8c4731d
feat: paper.bib formatting
