+++ paper.bib
@@ -1,46 +1,44 @@
 @misc{mediapipe,
-  author = {google-ai-edge},
-  title = {mediapipe: Cross-platform, customizable ML solutions for live and streaming media.},
-  year = {2019},
+  author    = {google-ai-edge},
+  title     = {mediapipe: Cross-platform, customizable ML solutions for live and streaming media.},
+  year      = {2019},
   publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/google-ai-edge/mediapipe}
+  journal   = {GitHub repository},
+  url       = {https://github.com/google-ai-edge/mediapipe}
 }
 
-@misc{mediapipe-hand‑landmarker,
+@misc{mediapipe-hand-landmarker,
+  author       = {google-ai-edge},
   title        = {MediaPipe Hand Landmarker (Hand Landmarks Detection Guide)},
   howpublished = {Google AI Edge},
-  note         = {Last updated January 13, 2025},
+  year         = {2019},
+  note         = {Last updated January 13, 2025},
   url          = {https://ai.google.dev/edge/mediapipe/solutions/vision/hand_landmarker}
 }
 
 @inproceedings{50649,
-  title	= {GHUM & GHUML: Generative 3D Human Shape and Articulated Pose Models},
-  author	= {Hongyi Xu and Eduard Gabriel Bazavan and Andrei Zanfir and Bill Freeman and Rahul Sukthankar and Cristian Sminchisescu},
-  year	= {2020},
-  URL	= {https://openaccess.thecvf.com/content_CVPR_2020/html/Xu_GHUM__GHUML_Generative_3D_Human_Shape_and_Articulated_Pose_CVPR_2020_paper.html},
-  booktitle	= {IEEE/CVF Conference on Computer Vision and Pattern Recognition (Oral)},
-  pages	= {6184-6193}
+  title     = {GHUM & GHUML: Generative 3D Human Shape and Articulated Pose Models},
+  author    = {Hongyi Xu and Eduard Gabriel Bazavan and Andrei Zanfir and Bill Freeman and Rahul Sukthankar and Cristian Sminchisescu},
+  year      = {2020},
+  url       = {https://openaccess.thecvf.com/content_CVPR_2020/html/Xu_GHUM__GHUML_Generative_3D_Human_Shape_and_Articulated_Pose_CVPR_2020_paper.html},
+  booktitle = {IEEE/CVF Conference on Computer Vision and Pattern Recognition (Oral)},
+  pages     = {6184-6193}
 }
 
 @misc{boddu_sri_pavan_2025_14632005,
-  author       = {Boddu Sri Pavan and
-                  Chandrasheker Thummanagoti and
-                  Boddu Swathi Sree},
-  title        = {KitikiPlot A Python library to visualize
-                   categorical sliding window data
-                  },
-  year         = 2025,
-  publisher    = {Zenodo},
-  doi          = {10.5281/zenodo.14632005},
-  url          = {https://doi.org/10.5281/zenodo.14632005},
+  author    = {Boddu Sri Pavan and Chandrasheker Thummanagoti and Boddu Swathi Sree},
+  title     = {KitikiPlot A Python library to visualize categorical sliding window data},
+  year      = {2025},
+  publisher = {Zenodo},
+  doi       = {10.5281/zenodo.14632005},
+  url       = {https://doi.org/10.5281/zenodo.14632005}
 }
 
 @misc{visualise-data-kit,
-  author = {Nagendra Dharmireddi and Jhansi Pyla},
-  title = {visualise-data-kit: Skeleton based visualizations of Google's MediaPipe},
-  year = {2024},
+  author    = {Nagendra Dharmireddi and Jhansi Pyla},
[…]
