### 2025-07-29 #8394 @BodduSriPavan-111
Updated ```paper.bib``` as suggested. Running check
