### 2025-07-29 learn-hunger/visual-gestures#4 @boisgera
However the bibliography items do not appear in the paper. I guess that you need to explicitly [cite every entry](https://joss.readthedocs.io/en/latest/paper.html#citations) that you want to appear a the end of the paper (after a "References" section title that you have to add manually).
