### 2025-07-29 f789ea9
feat: included-references-section
Added 'References' section
