### 2025-07-29 b21f312
feat: included-references-manually
Included references manually in 'paper.bib'
