### 2025-09-08 learn-hunger/visual-gestures#9 @boisgera
Statement of Need
JOSS Context: https://github.com/openjournals/joss-reviews/issues/8394

Wrt the current paper Statement of need section:

> Traditional cursor control relies on touch-based interaction, requiring physical contact with a device. Transitioning to gesture-based control enhances user experience across various domains, including AR, VR, gaming, healthcare, e-commerce, and industrial automation.
>
> visual-gestures.js addresses the need for a ready-to-integrate, open-source solution, offering seamless integration, offline functionality, and high customizability, including debugger support for rapid development.

This text motivates properly introduces the need for gesture-based control but not any software toolkit that may already exist to manage it and how this project specifically differentiates from these options. This point needs to be developed.
