### 2025-09-08 2f004ad
docs: added-inline-refs-removed-explicit-refs
