+++ paper.bib
@@ -1,3 +1,14 @@
+@article{opencv_library,
+    author = {Bradski, G.},
+    citeulike-article-id = {2236121},
+    journal = {Dr. Dobb's Journal of Software Tools},
+    keywords = {bibtex-import},
+    posted-at = {2008-01-15 19:21:54},
+    priority = {4},
+    title = {{The OpenCV Library}},
+    year = {2000}
+}
+
 @misc{mediapipe,
   author    = {google-ai-edge},
   title     = {mediapipe: Cross-platform, customizable ML solutions for live and streaming media.},
