### 2025-09-09 learn-hunger/visual-gestures#9 @BodduSriPavan-111
@boisgera
Added the statement describing how the current project differentiates from the existing options.
