+++ src/app/pointer/README.md
@@ -1,0 +1,222 @@
+# Custom Events – Low-Level Value Definitions for Customization
+
+This directory contains definitions of custom low-level values used by **visual-gestures.js** 
+to detect and interpret finger movements. These values are computed directly from hand landmarks
+returned by the hand-tracking model.
+
+The purpose of this document is to make these formulas explicit so that users can understand and, 
+if needed, customize or extend them.
+
+## palmHeight
+- Represents vertical size of human palm.  
+- Defined as the "Weighted Euclidean distance" between wrist landmark and index finger MCP.  
+
+**Formula:**
+
+$$
+\text{palmHeight} = d_w(\text{WRIST}, \text{INDEX.MCP}, \mathbf{w})
+$$
+where  
+$$
+d_w(p, q, \mathbf{w}) = \sqrt{ \sum_{i=1}^n w_i (p_i - q_i)^2 }
+$$
+and the weight vector is  
+$$
+\mathbf{w} = (w_x, w_y) = (0.5, 1.0)
+$$
+
+**Rationale:**  
+The y-axis is weighted more heavily than the x-axis to better reflect real-world scenarios where 
+vertical variation dominates palm height estimation.
+
+
+
+## fingerHeight
+- Represents the vertical length of a finger.
+- Defined as the "Weighted Euclidean Distance" between the tip of the finger (here Index finger TIP) and its MCP.  
+
+**Formula:** 
+
+$$
+\text{fingerHeight} = d_w(\text{INDEX.TIP}, \text{INDEX.MCP}, \mathbf{w})
+$$
+where  
+$$
+d_w(p, q, \mathbf{w}) = \sqrt{ \sum_{i=1}^n w_i (p_i - q_i)^2 }
+$$
+and the weight vector is  
+$$
+\mathbf{w} = (w_x, w_y) = (0.5, 1.0)
+$$
+
+**Rationale:**  
+Similar to **palmHeight**, the y-axis is weighted more heavily than the x-axis, 
+giving higher importance to vertical displacement when estimating finger length.
+
+
+## fingerKinkRatio
+
+The **fingerKinkRatio** is a novel metric designed to determine whether a human finger is 
+folded or extended.  
+
+**Formula:**  
+
+$$
+\text{fingerKinkRatio} = \frac{1000 \times \text{fingerHeight}}{\text{palmHeight}}
+$$
+
+**Properties:**  
+1. Indicates the degree of finger folding/unfolding.   
+2. Independent of the camera–hand distance (scale-invariant).  
+3. Sensitive only to finger motion (closing/opening).  
+4. Multiplication by 1000 normalizes the ratio into a convenient range (approximately **1–1000**).  
+
+## MCPtoTIPDistance
+
+Measures the weighted Euclidean distance between the MCP joint and the TIP of a finger.  
+
+**Formula:**  
+
+$$
+d_{\text{MCP→TIP}}(f) = \sqrt{ w_x \cdot (x_{\text{TIP}} - x_{\text{MCP}})^2 + w_y \cdot (y_{\text{TIP}} - y_{\text{MCP}})^2 }
+$$  
+
+where for this implementation:  
+- \( w_x = 0.5 \)  
+- \( w_y = 1 \)  
+- \( f \) = chosen finger (e.g., INDEX, MIDDLE, RING, PINKY, or THUMB)  
+
+**Properties:**  
+1. Provides a normalized distance measure accounting for hand pose.  
+2. Y-axis is given higher weight (\(w_y > w_x\)) to better reflect real-world vertical hand motion.
+
+## piecewiseDistance
+
+- Computes the sum of weighted Euclidean distances between consecutive joints of a finger.
[…]
