### 2025-09-16 learn-hunger/visual-gestures#9 @boisgera
The current "Statement of need" section:

> Traditional cursor control relies on touch-based interaction, requiring physical contact with a device. Transitioning to gesture-based control enhances user experience across various domains, including AR, VR, gaming, healthcare, e-commerce, and industrial automation.

> While MediaPipe @50649 @mediapipe @mediapipe-hand-landmarker and OpenCV @opencv_library offer strong hand-tracking, they are not ready-to-use cursor control toolkits. visual-gestures.js bridges this gap by providing a ready-to-integrate, open-source solution, offering seamless integration, offline functionality, and high customizability, including debugger support for rapid development.

  - I'd appreciate more details in the second paragraph; what makes visual-gestures ready-to-use when OpenCv and MediaPipe are not? What do you mean by "high customizability", etc. Some elements that would be more factual/"refutable" thatn the current description (and would sound less like an advertisment for the project).

  - About the first paragraph. I have been discussing @IsmailYC and @alexispaljic about the AR/VR case (thanks for the feedback! 🙏). They are indeed using gestures a lot but the common VR headsets (Meta Quest, etc.) incorporate their own cameras and detection software to produce the relevant hand skeleton information, which I guess would make the use of your project difficult if not impossible to use. While you do specifically say that your project specificically is relevant in this context, it is implicit. Which begs the question: for which application context would your project specifically be relevant and why? Is there any experience of if being effectively used in a specific domain (healthcare, e-commerce, etc.)?
