+++ paper.md
@@ -46,15 +46,20 @@ available on GitHub \: https://github.com/learn-hunger/visual-gestures
 # Statement of need
 
 Traditional cursor control relies on touch-based interaction, requiring 
-physical contact with a device. Transitioning to gesture-based control 
-enhances user experience across various domains, including AR, VR, gaming, 
-healthcare, e-commerce, and industrial automation.
+physical contact with a device. While libraries such as MediaPipe @50649 @mediapipe @mediapipe-hand-landmarker 
+and OpenCV @opencv_library provide accurate hand landmarks, they leave developers with raw coordinate data. 
+Turning these into usable interactions still requires custom rule-based logic and event handling.
 
-While MediaPipe @50649 @mediapipe @mediapipe-hand-landmarker and OpenCV @opencv_library offer strong hand-tracking,
-they are not ready-to-use cursor control toolkits. **visual-gestures.js** bridges this gap 
-by providing a ready-to-integrate, open-source solution, offering seamless integration, 
-offline functionality, and high customizability, including debugger support for 
-rapid development.
+visual-gestures.js fills this gap by:
+- Performing the rule-based classification of landmarks into gestures like click, drag, and hover.
+- Exposing these gestures through an API that mimics native browser pointer events (pointerMove, pointerDown, pointerUp),
+  making integration seamless.
+- Providing customizable gesture definitions with adjustable landmarks, thresholds, and motion rules, combined with lightweight
+  client-side formulas for smoothing and robust detection, optimized for real-time use in the browser.
+- Enabling touchless websites and WebXR prototypes using only a webcam, without servers or proprietary SDKs.
+
+This combination of rule-based classification and event abstraction makes visual-gestures.js a practical, 
+ready-to-use toolkit in websites that complements, rather than duplicates, existing hand-tracking frameworks.
 
 # Proposed Methodology
 
