### 2025-09-19 learn-hunger/visual-gestures#9 @BodduSriPavan-111
@boisgera
Updated the "Statement of need" as suggested.
