### 2025-11-13 #8394 @boisgera
Hi @chand-lab

> Yes. Sorry! I have other grants due and have been slow.

Sorry for nagging you again ... Do you think that you will have some time shortly to review the submission?

Best regards,

Sébastien Boisgérault
JOSS editor
