### 2025-11-13 #8394 @chand-lab
Yes! Sorry I was away in India for a few weeks and my son destroyed my
laptop. I will do it this Friday.

Best,

Chand

Chandramouli Chandrasekaran
Assistant Professor of Anatomy and Neurobiology
Assistant Professor of Psychological and Brain Sciences
Boston University
Boston, MA, USA
http://www.chandlab.org
Ph: 617 358 8359 (MED)

On Thu, Nov 13, 2025 at 4:18 AM Sébastien Boisgérault <
***@***.***> wrote:

> *boisgera* left a comment (openjournals/joss-reviews#8394)
> <https://github.com/openjournals/joss-reviews/issues/8394#issuecomment-3526674587>
>
> Hi @chand-lab <https://github.com/chand-lab>
>
> Yes. Sorry! I have other grants due and have been slow.
>
> Sorry for nagging you again ... Do you think that you will have some time
> shortly to review the submission?
>
> Best regards,
>
> Sébastien Boisgérault
> JOSS editor
>
