### 2026-01-03 #8394 @BodduSriPavan-111
@crvernon Thank you for your patience throughout this review process and for bringing these important concerns to our attention. We'd like to address each point carefully:

## 1. Reusability and Extensibility

**Reusability:** Our package architecture is designed with modularity in mind. Core components are structured as reusable modules that can be independently integrated:
1. The event handling system is abstracted into reusable APIs that developers can leverage to create custom gesture events beyond our predefined set.
2. Novel algorithms including "Fluid Kink Ratio (finger bend detection)", Hibernating Sliding Window (temporal gesture analysis), and our weighted Euclidean distance decision-making algorithm are all exposed as reusable functions.
3. These algorithms can be applied to different finger open-close analysis and gesture recognition scenarios beyond cursor control.

**Extensibility:** The package provides clear extension points:
1. Developers can easily integrate visual-gestures with existing JavaScript event handlers through our event system.
2. The landmark-based detection system allows customization of which landmarks serve as control points.
3. New gesture events can be created using our existing API infrastructure without modifying core code.

## 2. Scope Alignment with JOSS Guidelines

We submitted on February 15, 2025, referencing the JOSS submission guidelines available at that time (https://github.com/openjournals/joss/blob/cbd745c740c6abbc17ce525ee498d83817c2344d/docs/submitting.md). According to those guidelines, web-based tools are in scope if they:
> "are built around and expose a 'core library' through a web-based experience (e.g., R/Shiny applications)"

Our submission aligns with criterion (1) as per submission guidelines:
While we utilize "@mediapipe/tasks-vision" for hand landmark extraction, this is analogous to how R/Shiny applications use R as their computational backend.
Our novel scientific contribution lies in the algorithms built on top of these landmarks:
1. Fluid Kink Ratio formula for finger bend quantification.
2. Hibernating Sliding Window, and Weighted Euclidean distance method for temporal gesture state management.
3. Event mapping architecture that translates hand gestures to standardized DOM events.

These algorithms and methods constitute the core library that we expose through our npm package. The GUI/cursor control functionality is the demonstration application, not the primary contribution. The package is:
1. Platform-independent (works offline, no internet required).
2. Fully testable locally (as documented in our repository).
3. Available as a reusable TypeScript/JavaScript library via npm.
4. Accompanied by comprehensive documentation, demo, and examples.

The submission adheres to the provided guidelines. While the work is not limited to GUI aspects, we did not find any explicit mention of GUI-related constraints in the submission guidelines.

## 3. AI Usage in Development

No AI code generation was used in developing our codebase. All algorithms, architectural decisions, and implementations were developed manually by our team.

We believe visual-gestures offers genuine scientific and software engineering contributions through our novel gesture recognition algorithms and reusable architecture. We're happy to provide additional clarification, documentation, or make any adjustments needed to demonstrate our package meets JOSS requirements.

Thank you for your consideration.

cc @boisgera
