### 2026-01-16 #8394 @crvernon
@BodduSriPavan-111 - after further review with our larger editorial board, we must reject this submission at this time.  I do appreciate you taking time to provide feedback; however, JOSS editors noted that this was not deemed as research software with an obvious research application.  Also, it was pointed out that there were no tests and there we still major concerns about reusability.

We do wish you all the best in your future open source development.

cc:  @boisgera
